Adding uel tests does a bin diff of GDX files written by wrapper and port.

+++ tests/gdxinterfacetests.cpp
@@ -2,6 +2,9 @@
 #include "doctest.h"
 #include "../xpwrap.h"
 #include "../gxfile.h"
+#include "../utils.h"
+
+using namespace std::literals::string_literals;
 
 namespace tests::gdxinterfacetests {
     TEST_SUITE_BEGIN("GDX interface tests");
@@ -27,23 +30,42 @@ namespace tests::gdxinterfacetests {
         pgx.gdxClose();
     }
 
-    template<typename T> void runAddingUELTest() {
+    enum UELTestOps {
+        read = 0x01,
+        write = 0x10,
+        readAndWrite = 0x11
+    };
+
+    template<typename T> void runAddingUELTest(const std::string &fn, UELTestOps ops, bool cleanup=false) {
         int ErrNr;
-        std::string ErrMsg, fn {"gdxinterfacetest.gdx"};
+        std::string ErrMsg;
+        if(ops & write)
         {
             auto pgxObj {std::make_unique<T>(ErrMsg)};
             writeFileWithSingleUEL(*pgxObj, fn, ErrNr);
         }
+        if(ops & read)
         {
             auto pgxObj {std::make_unique<T>(ErrMsg)};
             readFileWithSingleUEL(*pgxObj, fn, ErrNr);
         }
-        std::filesystem::remove(fn);
+        if(cleanup)
+            std::filesystem::remove(fn);
     };
 
     TEST_CASE("Test adding uels") {
-        runAddingUELTest<xpwrap::GDXFile>();
-        //runAddingUELTest<gxfile::TGXFileObj>();
+        std::array filenames{ "uel_wrapper.gdx"s, "uel_port.gdx"s };
+        runAddingUELTest<xpwrap::GDXFile>(filenames.front(), write);
+        runAddingUELTest<gxfile::TGXFileObj>(filenames.back(), write);
+        auto maybeMismatches = utils::binaryFileDiff(filenames.front(), filenames.back());
+        if (maybeMismatches) {
+            for (const auto& mm : *maybeMismatches)
+                std::cout << "Mismatch at offset " << mm.offset << " with left hand side = " << mm.lhs << " and right hand side = " << mm.rhs << "\n";
+            std::cout << std::endl;
+        }
+        else std::cout << "No mismatches found!" << std::endl;
+        runAddingUELTest<xpwrap::GDXFile>(filenames.front(), read, true);
+        runAddingUELTest<gxfile::TGXFileObj>(filenames.back(), read, true);
     }
 
     TEST_SUITE_END();
